refactor: split kernel postgres adapter modules
