+++ packages/kernel-adapter-postgres/CHANGELOG.md
@@ -2,6 +2,7 @@ English | [中文文档](CHANGELOG.zh-CN.md)
 
 ## [Unreleased]
 
+- refactor(package): split the Postgres repository implementation into semantic repository, factory, schema, mapper, interface, type, and utility modules without widening the public API.
 - chore(package): adopt the kernel-adapter-postgres package name to keep it scoped to the kernel Postgres repository adapter.
 - chore(package): add the kernel-adapter-postgres changelog baseline for the Postgres MetaKernel repository implementation package.
