+++ packages/kernel-adapter-postgres/CHANGELOG.zh-CN.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 
+- refactor(package): 将 Postgres repository 实现拆成 repository、factory、schema、mapper、interface、type 与 utility 模块，同时不扩大 public API。
 - chore(package): 采用 kernel-adapter-postgres 包名，明确它只承载 kernel Postgres repository adapter。
 - chore(package): 为承载 Postgres MetaKernel repository 实现的 kernel-adapter-postgres 包补充 changelog 基线。
