+++ packages/kernel-adapter-postgres/src/postgres/factory/postgres-meta-kernel-repository.factory.ts
@@ -1,0 +1,6 @@
+import type { DbConfig } from "@zhongmiao/meta-lc-kernel";
+import { PostgresMetaKernelRepository } from "../repository/postgres-meta-kernel.repository";
+
+export function createPostgresMetaKernelRepository(config: DbConfig): PostgresMetaKernelRepository {
+  return new PostgresMetaKernelRepository(config);
+}
