+++ packages/kernel-adapter-postgres/src/postgres/index.ts
@@ -1,1 +1,3 @@
-export * from "./postgres-meta-kernel-repository";
+export { PostgresMetaKernelRepository } from "./repository/postgres-meta-kernel.repository";
+export { createPostgresMetaKernelRepository } from "./factory/postgres-meta-kernel-repository.factory";
+export type { MigrationExecutionContext } from "./interfaces";
