+++ packages/kernel-adapter-postgres/src/postgres/interfaces/index.ts
@@ -1,0 +1,3 @@
+export type { MigrationExecutionContext } from "./migration-execution-context.interface";
+export type { MetaVersionRow } from "./meta-version-row.interface";
+export type { MetaDefinitionVersionRow } from "./meta-definition-version-row.interface";
