+++ packages/kernel-adapter-postgres/src/postgres/interfaces/migration-execution-context.interface.ts
@@ -1,0 +1,6 @@
+export interface MigrationExecutionContext {
+  appId: string;
+  fromVersion: number;
+  toVersion: number;
+  requestId?: string;
+}
