+++ packages/kernel-adapter-postgres/src/postgres/types/index.ts
@@ -1,0 +1,1 @@
+export type { MigrationAuditStatus } from "./migration-audit-status.type";
