+++ packages/kernel-adapter-postgres/src/postgres/types/migration-audit-status.type.ts
@@ -1,0 +1,1 @@
+export type MigrationAuditStatus = "success" | "failure" | "blocked";
